# allow accumulation of merge operands before merge