--- a/compaction_iter.go
+++ b/compaction_iter.go
@@ -131,8 +131,6 @@ type compactionIter struct {
 	// determine when iteration has advanced to a new user key and thus a new
 	// snapshot stripe.
 	keyBuf []byte
-	// Temporary buffer used for aggregating merge operations.
-	valueBuf []byte
 	// Is the current entry valid?
 	valid     bool
 	iterKey   *InternalKey
@@ -257,7 +255,18 @@ func (i *compactionIter) Next() (*InternalKey, []byte) {
 			// invalidating the state that maybeZeroSeqnum uses to make its
 			// determination.
 			i.maybeZeroSeqnum()
-			return i.mergeNext()
+			var valueMerger ValueMerger
+			valueMerger, i.err = i.merge(i.iterKey.UserKey, i.iterValue)
+			if i.err == nil {
+				i.mergeNext(valueMerger)
+			}
+			if i.err == nil {
+				i.value, i.err = valueMerger.Finish()
+			}
+			if i.err == nil {
+				return &i.key, i.value
+			}
+			return nil, nil
 
 		case InternalKeyKindRangeDelete:
 			i.nextInStripe()
@@ -267,7 +276,8 @@ func (i *compactionIter) Next() (*InternalKey, []byte) {
 			// NB: Invalid keys occur when there is some error parsing the key. Pass
 			// them through unmodified.
 			i.saveKey()
-			i.saveValue()
+			i.value = make([]byte, len(i.iterValue))
+			copy(i.value, i.iterValue)
 			i.iterNext()
 			i.valid = true
 			return &i.key, i.value
@@ -343,10 +353,9 @@ func (i *compactionIter) nextInStripe() bool {
 	return false
 }
 
-func (i *compactionIter) mergeNext() (*InternalKey, []byte) {
-	// Save the current key and value.
+func (i *compactionIter) mergeNext(valueMerger ValueMerger) {
+	// Save the current key.
 	i.saveKey()
-	i.saveValue()
 	i.valid = true
 
 	// Loop looking for older values in the current snapshot stripe and merge
@@ -354,52 +363,55 @@ func (i *compactionIter) mergeNext() (*InternalKey, []byte) {
 	for {
 		if !i.nextInStripe() {
 			i.skip = false
-			return &i.key, i.value
+			return
 		}
 		key := i.iterKey
 		switch key.Kind() {
 		case InternalKeyKindDelete:
 			// We've hit a deletion tombstone. Return everything up to this point and
 			// then skip entries until the next snapshot stripe.
-			i.valueBuf = i.value[:0]
 			i.skip = true
-			return &i.key, i.value
+			return
 
 		case InternalKeyKindRangeDelete:
 			// We've hit a range deletion tombstone. Return everything up to this
 			// point and then skip entries until the next snapshot stripe.
 			i.skip = true
-			return &i.key, i.value
+			return
 
 		case InternalKeyKindSet:
 			if i.rangeDelFrag.Deleted(*key, i.curSnapshotSeqNum) {
 				i.skip = true
-				return &i.key, i.value
+				return
 			}
 
 			// We've hit a Set value. Merge with the existing value and return. We
 			// change the kind of the resulting key to a Set so that it shadows keys
 			// in lower levels. That is, MERGE+MERGE+SET -> SET.
-			i.value = i.merge(i.key.UserKey, i.iterValue, i.value, nil)
-			i.valueBuf = i.value[:0]
+			i.err = valueMerger.MergeOlder(i.iterValue)
+			if i.err != nil {
+				return
+			}
 			i.key.SetKind(InternalKeyKindSet)
 			i.skip = true
-			return &i.key, i.value
+			return
 
 		case InternalKeyKindMerge:
 			if i.rangeDelFrag.Deleted(*key, i.curSnapshotSeqNum) {
 				i.skip = true
-				return &i.key, i.value
+				return
 			}
 
 			// We've hit another Merge value. Merge with the existing value and
 			// continue looping.
-			i.value = i.merge(i.key.UserKey, i.iterValue, i.value, nil)
-			i.valueBuf = i.value[:0]
+			i.err = valueMerger.MergeOlder(i.iterValue)
+			if i.err != nil {
+				return
+			}
 
 		default:
 			i.err = fmt.Errorf("invalid internal key kind: %d", i.iterKey.Kind())
-			return nil, nil
+			return
 		}
 	}
 }
@@ -446,11 +458,6 @@ func (i *compactionIter) singleDeleteNext() bool {
 func (i *compactionIter) saveKey() {
 	i.keyBuf = append(i.keyBuf[:0], i.iterKey.UserKey...)
 	i.key.UserKey = i.keyBuf
-}
-
-func (i *compactionIter) saveValue() {
-	i.valueBuf = append(i.valueBuf[:0], i.iterValue...)
-	i.value = i.valueBuf
 }
 
 func (i *compactionIter) cloneKey(key InternalKey) InternalKey {