Changed the merge operator API to support adding operands one by one via
`Merge{Newer,Older}()` followed by a call to `Finish()`. This gives
implementations the option to accumulate merge operands, only doing the
final merge work when `Finish()` is called. For example, it can be used
to reduce (de)serialization of partial merge results compared to the
previous binary operation API.