@ajkr 2019-10-28
Here are the results of my prototype of using this new API together with the fixed benchmark change in https://github.com/cockroachdb/cockroach/pull/41966.

```
name                                                    old time/op  new time/op  delta
MVCCGetMergedTimeSeries/numKeys=1/mergesPerKey=1-6      6.42µs ± 3%  5.13µs ± 3%  -20.10%  (p=0.000 n=10+9)
MVCCGetMergedTimeSeries/numKeys=1/mergesPerKey=16-6     63.9µs ± 2%  40.7µs ± 3%  -36.20%  (p=0.000 n=9+9)
MVCCGetMergedTimeSeries/numKeys=1/mergesPerKey=256-6    9.80ms ± 3%  0.91ms ± 2%  -90.72%  (p=0.000 n=10+9)
MVCCGetMergedTimeSeries/numKeys=16/mergesPerKey=1-6     6.71µs ± 2%  8.53µs ± 2%  +27.02%  (p=0.000 n=9+9)
MVCCGetMergedTimeSeries/numKeys=16/mergesPerKey=16-6    65.7µs ± 3%  77.2µs ± 2%  +17.47%  (p=0.000 n=10+10)
MVCCGetMergedTimeSeries/numKeys=16/mergesPerKey=256-6   9.85ms ± 2%  1.81ms ± 4%  -81.61%  (p=0.000 n=10+10)
MVCCGetMergedTimeSeries/numKeys=256/mergesPerKey=1-6    7.15µs ± 2%  9.13µs ± 3%  +27.70%  (p=0.000 n=10+10)
MVCCGetMergedTimeSeries/numKeys=256/mergesPerKey=16-6   65.0µs ± 2%  79.9µs ± 2%  +22.98%  (p=0.000 n=10+10)
MVCCGetMergedTimeSeries/numKeys=256/mergesPerKey=256-6  9.90ms ± 3%  1.80ms ± 2%  -81.82%  (p=0.000 n=10+8)
```

One might wonder why the single operand case got worse compared to the benchmark fix alone. It's because after this PR Pebble actually pushes single merge operands through the merge operator which is necessary for correctness as they may be unsorted.