@petermattis 2019-10-28
Are these Pebble (old) vs Pebble (new) numbers? Or RocksDB (old) vs Pebble (new)?

Why did `numKeys=16/mergesPerKey=16` and `numKeys=256/mergesPerKey=16` get slower? I would have expected those to get faster similar to the `mergesPerKey=256` runs.